@@ -1,28 +1,48 @@
 {
-  "name": "childsafeai-presentation",
+  "name": "covid-catchphrase",
   "version": "0.0.1",
   "engines": {
-    "node": "0.12.x"
+    "node": "13.13.x"
   },
   "dependencies": {
     "@tailwindcss/ui": "^0.1.3",
-    "alpinejs": "^2.2.5",
+    "babelify": "^10.0.0",
     "bower": "^1.4.1",
-    "gulp": "^3.9.0",
+    "browserify": "^16.5.1",
+    "gulp": "^4.0.x",
     "gulp-bro": "^2.0.0",
+    "gulp-cli": "^2.2.0",
     "gulp-jshint": "^1.11.2",
-    "gulp-livereload": "^3.8.0",
-    "gulp-open": "^0.3.2",
+    "gulp-livereload": "^4.0.2",
+    "gulp-open": "^3.0.1",
     "gulp-postcss": "^8.0.0",
-    "gulp-sass": "^2.0.3",
+    "gulp-sass": "^4.0.2",
     "gulp-sourcemaps": "^2.6.5",
-    "gulp-watch": "^4.2.4",
+    "gulp-vueify2": "0.0.3",
+    "gulp-watch": "^5.0.1",
     "gulp-webserver": "^0.9.1",
     "jshint-stylish": "^2.0.1",
     "tailwindcss": "^1.2.0",
+    "vue": "^2.6.11",
+    "vue-smooth-dnd": "^0.8.1",
+    "vueify-next": "^9.5.3",
+    "vuex": "^3.2.0",
+    "vuex-persist": "^2.2.0",
     "yargs": "^3.15.0"
   },
   "scripts": {
     "postinstall": "bower install"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.9.5",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/preset-stage-2": "^7.8.3",
+    "@babel/preset-stage-3": "^7.8.3",
+    "@vue/cli-service": "^4.3.1",
+    "vue-template-compiler": "^2.6.11"
   }
 }
