@@ -1,2 +1,7 @@
-'use strict'
-let Alpine = require('alpinejs');
+import Vue from 'vue'
+import Vuex from 'vuex'
+import App from './App.vue'
+
+new Vue({
+  render: h => h(App)
+}).$mount('#app')
