@@ -1,6 +1,13 @@
 // tailwind.config.js
 module.exports = {
+  theme: {
+    extend: {
+      colors: {
+        green: "#00bc70"
+      }
+    }
+  },
   plugins: [
     require('@tailwindcss/ui'),
-  ]
-}
+  ],
+};
